Refactor database model imports and restructure document handling
- Introduced new models for security, including BannedSubnet, LoginThrottle, and TrafficThrottle.
- Updated import paths for various models across the codebase to reflect the new database structure.
- Removed legacy model paths and ensured that all references to models are consistent with the new organization.
- Deleted unused document base and metadata files to streamline the codebase.
- Added tests to verify the registration of core database tables and ensure legacy paths are no longer accessible.
