+++ src/alembic/env.py
@@ -18,28 +18,9 @@ if config.config_file_name is not None:
 # for 'autogenerate' support
 # from myapp import mymodel
 # target_metadata = mymodel.Base.metadata
+import include.database.models  # noqa: F401
+
 from include.database.session import Base
-from include.domains.access.models import UserBlockEntry, UserBlockSubEntry
-from include.domains.identity.models import (
-    User,
-    UserGroup,
-    UserGroupPermission,
-    UserPermission,
-    UserMembership,
-)
-from include.domains.documents.files import File, FileTask
-from include.domains.documents import (
-    BaseObject,
-    Document,
-    DocumentRevision,
-    DocumentAccessRule,
-    DocumentMetadata,
-    DocumentMetadataTag,
-    Folder,
-    FolderAccessRule,
-)
-from include.domains.keyrings.models import UserKey
-from include.domains.security import BannedSubnet, LoginThrottle, TrafficThrottle
 
 target_metadata = Base.metadata
