+++ src/include/database/models/access.py
@@ -7,7 +7,7 @@ from sqlalchemy.orm import Mapped, mapped_column, relationship
 from include.database.session import Base
 
 if TYPE_CHECKING:
-    from include.domains.identity.models import User
+    from include.database.models.identity import User
 
 
 class UserBlockEntry(Base):
