+++ src/include/database/session.py
@@ -77,7 +77,7 @@ def _add_filtering_criteria(execute_state: ORMExecuteState) -> None:
         and not execute_state.is_column_load
         and not execute_state.execution_options.get("include_deleted", False)
     ):
-        from include.domains.documents import Document, EntityStatus, Folder
+        from include.database.models.documents import Document, EntityStatus, Folder
 
         execute_state.statement = execute_state.statement.options(
             with_loader_criteria(Folder, Folder.status != EntityStatus.DELETED),
