+++ src/include/domains/documents/handlers/revisions.py
@@ -1,11 +1,11 @@
+from include.database.models.documents import Document, DocumentRevision
+from include.database.models.identity import User
 from include.database.session import Session
 from include.domains.access.permissions import Permissions
-from include.domains.documents import Document, DocumentRevision
 from include.domains.documents.handlers.documents import (
     create_file_task,
     mark_document_modified,
 )
-from include.domains.identity.models import User
 from include.domains.operations.messages import Messages as smsg
 from include.transport.connection import ConnectionHandler
 from include.transport.request_handler import RequestHandler
