+++ src/include/domains/identity/handlers/groups.py
@@ -3,15 +3,15 @@ __all__ = [
     # ...
 ]
 
-from include.database.session import Session
-from include.domains.access.permissions import Permissions
-from include.domains.identity.commands.groups import create_group
-from include.domains.identity.models import (
+from include.database.models.identity import (
     User,
     UserGroup,
     UserGroupPermission,
     UserMembership,
 )
+from include.database.session import Session
+from include.domains.access.permissions import Permissions
+from include.domains.identity.commands.groups import create_group
 from include.domains.operations.messages import Messages as smsg
 from include.transport.connection import ConnectionHandler
 from include.transport.request_handler import RequestHandler
