+++ src/include/domains/operations/handlers/system.py
@@ -3,12 +3,12 @@ import time
 import orjson
 from sqlalchemy import desc, func, true, update
 
+from include.database.models.files import FileTask
+from include.database.models.identity import User
+from include.database.models.operations import AuditEntry
 from include.database.session import Session
 from include.domains.access.permissions import Permissions
-from include.domains.documents.files import FileTask
-from include.domains.identity.models import User
 from include.domains.operations.messages import Messages as smsg
-from include.domains.operations.models import AuditEntry
 from include.shared import lockdown_enabled
 from include.transport.connection import ConnectionHandler
 from include.transport.request_handler import RequestHandler
