+++ src/include/domains/security/__init__.py
@@ -1,1 +1,1 @@
-from .models import *
+__all__: list[str] = []
