+++ src/include/domains/security/handlers/debugging.py
@@ -1,6 +1,6 @@
+from include.database.models.identity import User
 from include.database.session import Session
 from include.domains.access.permissions import Permissions
-from include.domains.identity.models import User
 from include.domains.operations.messages import Messages as smsg
 from include.transport.connection import ConnectionHandler
 from include.transport.request_handler import RequestHandler
